Move thin client tests to proper place
